dijstra

+++ src/incatch/DijkstraAlgorithm.java
@@ -23,7 +23,7 @@ public class DijkstraAlgorithm {
     private Set<Node> settledNodes;
     private Set<Node> unSettledNodes;
     private Map<Node, Node> predecessors;
-    private Map<Node, Integer> distance;
+    private Map<Node, Double> distance;
 
     public DijkstraAlgorithm(Graph graph) {
         // create a copy of the array so that we can operate on this array
@@ -34,9 +34,9 @@ public class DijkstraAlgorithm {
     public void execute(Node source) {
         settledNodes = new HashSet<Node>();
         unSettledNodes = new HashSet<Node>();
-        distance = new HashMap<Node, Integer>();
+        distance = new HashMap<Node, Double>();
         predecessors = new HashMap<Node, Node>();
-        distance.put(source, 0);
+        distance.put(source, 0.);
         unSettledNodes.add(source);
         while (unSettledNodes.size() > 0) {
             Node node = getMinimum(unSettledNodes);
@@ -60,7 +60,7 @@ public class DijkstraAlgorithm {
 
     }
 
-    private int getDistance(Node node, Node target) {
+    private double getDistance(Node node, Node target) {
         for (Edge edge : edges) {
             if (edge.getSource().equals(node)
                     && edge.getDestination().equals(target)) {
@@ -99,10 +99,10 @@ public class DijkstraAlgorithm {
         return settledNodes.contains(vertex);
     }
 
-    private int getShortestDistance(Node destination) {
-        Integer d = distance.get(destination);
+    private double getShortestDistance(Node destination) {
+        Double d = distance.get(destination);
         if (d == null) {
-            return Integer.MAX_VALUE;
+            return Double.MAX_VALUE;
         } else {
             return d;
         }
